+++ CHANGELOG.md
@@ -5,6 +5,7 @@ v4.0
 - [feat] timepicker - trigger `change` event only on click, not on mousemove
 - [feat] date-range picker - can be enabled by setting `isRange` property
 - [feat] slot[name="action-row"] added
+- [feat] add prop `disabledDatesFn` to disable dates. `{function(Date): boolean}`
 - [breaking] removed `positionFn` property
 - [breaking] removed `clearToggle` property (now this functionality - ability to clear the value is controlled by `required` property)
 - [breaking] method to register component as custom element has changed from `registerElement` to `registerSveltyPicker`. Also arguments has changed